+++ src/hash_table.rs
@@ -472,6 +472,36 @@ where
         Ok(post_processor(None))
     }
 
+    /// Finds any entry that satisfies the specified predicate.
+    #[inline]
+    fn any_entry<P: FnMut(&K, &V) -> bool>(&self, mut pred: P) -> bool {
+        let barrier = Barrier::new();
+        let mut current_array_ptr = self.bucket_array().load(Acquire, &barrier);
+        while let Some(current_array) = current_array_ptr.as_ref() {
+            self.clear_old_array(current_array, &barrier);
+            for index in 0..current_array.num_buckets() {
+                let bucket = current_array.bucket(index);
+                if let Some(locker) = Reader::lock(bucket, &barrier) {
+                    let data_block = current_array.data_block(index);
+                    let mut entry_ptr = EntryPtr::new(&barrier);
+                    while entry_ptr.next(*locker, &barrier) {
+                        let (k, v) = entry_ptr.get(data_block);
+                        if pred(k, v) {
+                            return true;
+                        }
+                    }
+                }
+            }
+
+            let new_current_array_ptr = self.bucket_array().load(Acquire, &barrier);
+            if current_array_ptr.without_tag() == new_current_array_ptr.without_tag() {
+                break;
+            }
+            current_array_ptr = new_current_array_ptr;
+        }
+        false
+    }
+
     /// Retains entries that satisfy the specified predicate.
     #[inline]
     fn retain_entries<F: FnMut(&K, &mut V) -> bool>(&self, mut pred: F) -> (usize, usize) {