+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 1.7.1
 
-* API update: add `hash_map::OccupiedEntry::{remove_and_next, remove_and_next_async}`.
+* Minor optimization.
 
 1.7.0